lv4

+++ LV4/2.zad/main.c
@@ -22,16 +22,16 @@
 /* fali zagrada nakon imenovanja funkcije main */
 int main(void) {
   
-  /* glupo postavljen zadatak pretpostavljam da a i b moraju bit int pa sam stavio int koji fali */
-  int a = 5;
-  int b = 7;
+  /* fali float ispred a i b iako smo mogli pisat i int pa radit typecasting kod racunanja al posto nije navedeno nis u zadatku stavio sam float */
+  float a = 5;
+  float b = 7;
   float c;
 
   /* ak nam je c float onda i a i b koje djelimo da bi dobili vrijednost c moraju bit float */
-  c = (float)a / (float)b;
+  c = a / b;
 
   /* nemoze ic %f jer ce nam ovaj izraz pored (c >= 0.1f && c <= 1.0f) vratit 1 ili 0 koji su int */
   printf("%d", c >= 0.1f && c <= 1.0f);
-
+  
   return 0;
 }
